requirements.txt
fastapi==0.120.1
uvicorn[standard]==0.38.0
python-jose[cryptography]==3.5.0
pydantic==2.12.3
pydantic-settings==2.11.0
httpx==0.28.1
pytest==8.4.2